+++ earth2studio/data/__init__.py
@@ -20,11 +20,11 @@ from .cbottle import CBottle3D
 from .cds import CDS
 from .cmip6 import CMIP6
 from .const import Constant, Constant_FX
+from .ecmwf import AIFS, AIFS_ENS, IFS, IFS_ENS
 from .gefs import GEFS_FX, GEFS_FX_721x1440
 from .gfs import GFS, GFS_FX
 from .goes import GOES
 from .hrrr import HRRR, HRRR_FX
-from .ifs import IFS
 from .imerg import IMERG
 from .jpss import JPSS
 from .ncar import NCAR_ERA5